backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.32
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
argon2-cffi>=23.1.0
pydantic-settings>=2.3.4
python-multipart>=0.0.9
celery>=5.4.0
redis>=5.0.7
httpx>=0.27.0
email-validator>=2.2.0
python-dateutil>=2.9.0.post0
orjson>=3.10.7
pytest>=8.3.2
croniter>=2.0.5
alembic>=1.14.0
loguru>=0.7.2
phonenumbers>=8.13.52
pymorphy3>=2.0.0
pymorphy3-dicts-ru>=2.4.0
anthropic>=0.39.0
